# movies_api_spec.yaml
openapi: 3.0.2
info:
  version: 2.0.0
  title: API de gestion de collection de films
  description: |
    API REST pour gérer une collection personnelle de films avec réalisateurs, genres et collections personnalisées.
    Cette API suit les principes HATEOAS (Hypermedia As The Engine Of Application State).
  contact:
    name: "Allegre Bastien"
    email: bastien.allegre@etu.univ-grenoble-alpes.fr
servers:
- url: http://localhost:3000/api
  description: Serveur de développement local
paths:
...
  /movies:
    get:
      tags:
        - Movies
      summary: Liste tous les films
      description: Récupère la liste complète des films avec pagination et filtres
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - name: status
          in: query
          description: Filtre par statut de visionnage
          schema:
            type: string
            enum: [a_voir, vu, en_cours]
        - name: genre_id
          in: query
          description: Filtre par identifiant de genre
          schema:
            type: string
        - name: director_id
          in: query
          description: Filtre par identifiant de réalisateur
          schema:
            type: string
        - name: collection_id
          in: query
          description: Filtre par identifiant de collection
          schema:
            type: string
        - name: sort
          in: query
          description: Champ de tri
          schema:
            type: string
            enum: [titre, annee, note, date_ajout]
            default: date_ajout
        - name: order
          in: query
          description: Ordre de tri
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Liste des films retournée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 127
                  page:
                    type: integer
                    example: 0
                  limit:
                    type: integer
                    example: 20
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieWithLinks'
                  _links:
                    $ref: '#/components/schemas/PaginationLinks'
    post:
      tags:
        - Movies
      summary: Ajoute un nouveau film
      description: Crée un nouveau film dans la collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        "201":
          description: Film créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieWithLinks'
        "400":
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/{id}:
    get:
      tags:
        - Movies
      summary: Récupère un film spécifique
      description: Retourne les détails complets d'un film avec ses relations
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        "200":
          description: Film retourné avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailWithLinks'
        "404":
          description: Film non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Movies
      summary: Met à jour un film
      description: Modifie les informations d'un film existant
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        "200":
          description: Film mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieWithLinks'
        "404":
          description: Film non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Movies
      summary: Supprime un film
      description: Retire un film de la collection
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        "204":
          description: Film supprimé avec succès
        "404":
          description: Film non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /movies/search:
    get:
      tags:
        - Movies
      summary: Recherche de films
      description: Recherche des films par titre
      parameters:
        - name: q
          in: query
          required: true
          description: Terme de recherche
          schema:
            type: string
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: Résultats de recherche
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieWithLinks'
                  _links:
                    type: object
                    properties:
                      self:
                        $ref: '#/components/schemas/Link'

  # ==================== COLLECTIONS ====================
...
  /collections/{id}:
    get:
      tags:
        - Collections
      summary: Récupère une collection
      description: Retourne les détails d'une collection avec ses films
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        "200":
          description: Collection retournée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetailWithLinks'
        "404":
          description: Collection non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Collections
      summary: Met à jour une collection
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionInput'
      responses:
        "200":
          description: Collection mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionWithLinks'
    delete:
      tags:
        - Collections
      summary: Supprime une collection
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        "204":
          description: Collection supprimée
        "404":
          description: Collection non trouvée
  /collections/{id}/movies:
    get:
      tags:
        - Collections
      summary: Liste les films d'une collection
      description: Récupère tous les films appartenant à une collection
      parameters:
        - $ref: '#/components/parameters/idParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: Liste des films de la collection
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieWithLinks'
                  _links:
                    type: object
                    properties:
                      self:
                        $ref: '#/components/schemas/Link'
                      collection:
                        $ref: '#/components/schemas/Link'
    post:
      tags:
        - Collections
      summary: Ajoute un film à la collection
      description: Associe un film existant à cette collection
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - movie_id
              properties:
                movie_id:
                  type: string
                  description: ID du film à ajouter
      responses:
        "200":
          description: Film ajouté à la collection
        "404":
          description: Collection ou film non trouvé
  /collections/{id}/movies/{movie_id}:
    delete:
      tags:
        - Collections
      summary: Retire un film de la collection
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la collection
          schema:
            type: string
        - name: movie_id
          in: path
          required: true
          description: ID du film à retirer
          schema:
            type: string
      responses:
        "204":
          description: Film retiré de la collection
        "404":
          description: Collection ou film non trouvé

  # ==================== GENRES ====================
...
  /genres/{id}/movies:
    get:
      tags:
        - Genres
      summary: Liste les films d'un genre
      description: Récupère tous les films appartenant à un genre
      parameters:
        - $ref: '#/components/parameters/idParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: Liste des films du genre
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieWithLinks'
                  _links:
                    type: object
                    properties:
                      self:
                        $ref: '#/components/schemas/Link'
                      genre:
                        $ref: '#/components/schemas/Link'

  # ==================== DIRECTORS ====================
...
  /directors/{id}/movies:
    get:
      tags:
        - Directors
      summary: Liste les films d'un réalisateur
      description: Récupère la filmographie complète d'un réalisateur
      parameters:
        - $ref: '#/components/parameters/idParam'
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
      responses:
        "200":
          description: Filmographie du réalisateur
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/MovieWithLinks'
                  _links:
                    type: object
                    properties:
                      self:
                        $ref: '#/components/schemas/Link'
                      director:
                        $ref: '#/components/schemas/Link'

  # ==================== STATS ====================
...
components:
  parameters:
    idParam:
      name: id
      in: path
      required: true
      description: Identifiant unique de la ressource
      schema:
        type: string
    pageParam:
      name: page
      in: query
      description: Numéro de page (commence à 0)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    limitParam:
      name: limit
      in: query
      description: Nombre d'éléments par page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    MovieInput:
      type: object
      required:
        - titre
        - annee
      properties:
        titre:
          type: string
          example: "Inception"
        annee:
          type: integer
          minimum: 1888
          example: 2010
        director_id:
          type: string
          description: ID du réalisateur
        genre_ids:
          type: array
          items:
            type: string
          description: Liste des IDs de genres
        duree:
          type: integer
          description: Durée en minutes
          example: 148
        synopsis:
          type: string
        statut:
          type: string
          enum: [a_voir, vu, en_cours]
          default: "a_voir"
        note:
          type: number
          format: float
          minimum: 0
          maximum: 10
        commentaire:
          type: string
        affiche_url:
          type: string
          format: uri
        tmdb_id:
          type: integer
        tags:
          type: array
          items:
            type: string
    Movie:
      allOf:
        - $ref: '#/components/schemas/MovieInput'
        - type: object
          required:
            - id
            - date_ajout
          properties:
            id:
              type: string
            date_ajout:
              type: string
              format: date-time
            date_visionnage:
              type: string
              format: date-time
    MovieWithLinks:
      allOf:
        - $ref: '#/components/schemas/Movie'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                director:
                  $ref: '#/components/schemas/Link'
                genres:
                  type: array
                  items:
                    $ref: '#/components/schemas/Link'
                collections:
                  type: array
                  items:
                    $ref: '#/components/schemas/Link'
    MovieDetailWithLinks:
      allOf:
        - $ref: '#/components/schemas/Movie'
        - type: object
          properties:
            director:
              $ref: '#/components/schemas/DirectorWithLinks'
            genres:
              type: array
              items:
                $ref: '#/components/schemas/GenreWithLinks'
            collections:
              type: array
              items:
                $ref: '#/components/schemas/CollectionWithLinks'
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                update:
                  $ref: '#/components/schemas/Link'
                delete:
                  $ref: '#/components/schemas/Link'

    # ==================== COLLECTION SCHEMAS ====================
    CollectionInput:
      type: object
      required:
        - nom
      properties:
        nom:
          type: string
          example: "Films d'horreur préférés"
        description:
          type: string
        is_public:
          type: boolean
          default: false
    Collection:
      allOf:
        - $ref: '#/components/schemas/CollectionInput'
        - type: object
          required:
            - id
            - date_creation
          properties:
            id:
              type: string
            date_creation:
              type: string
              format: date-time
            nombre_films:
              type: integer
    CollectionWithLinks:
      allOf:
        - $ref: '#/components/schemas/Collection'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                movies:
                  $ref: '#/components/schemas/Link'
    CollectionDetailWithLinks:
      allOf:
        - $ref: '#/components/schemas/Collection'
        - type: object
          properties:
            movies:
              type: array
              items:
                $ref: '#/components/schemas/MovieWithLinks'
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                update:
                  $ref: '#/components/schemas/Link'
                delete:
                  $ref: '#/components/schemas/Link'
                add_movie:
                  $ref: '#/components/schemas/Link'

    # ==================== GENRE SCHEMAS ====================
    GenreInput:
      type: object
      required:
        - nom
      properties:
        nom:
          type: string
          example: "Science-Fiction"
        description:
          type: string
    Genre:
      allOf:
        - $ref: '#/components/schemas/GenreInput'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
            nombre_films:
              type: integer
    GenreWithLinks:
      allOf:
        - $ref: '#/components/schemas/Genre'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                movies:
                  $ref: '#/components/schemas/Link'
...
    DirectorInput:
      type: object
      required:
        - nom
        - prenom
      properties:
        nom:
          type: string
          example: "Nolan"
        prenom:
          type: string
          example: "Christopher"
        date_naissance:
          type: string
          format: date
        nationalite:
          type: string
        biographie:
          type: string
        photo_url:
          type: string
          format: uri
    Director:
      allOf:
        - $ref: '#/components/schemas/DirectorInput'
        - type: object
          required:
            - id
          properties:
            id:
              type: string
            nombre_films:
              type: integer
    DirectorWithLinks:
      allOf:
        - $ref: '#/components/schemas/Director'
        - type: object
          properties:
            _links:
              type: object
              properties:
                self:
                  $ref: '#/components/schemas/Link'
                movies:
                  $ref: '#/components/schemas/Link'
...
    Link:
      type: object
      required:
        - href
      properties:
        href:
          type: string
          format: uri
          example: "/api/movies/123"
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH]
          default: "GET"
        rel:
          type: string
          example: "self"
    PaginationLinks:
      type: object
      properties:
        self:
          $ref: '#/components/schemas/Link'
        first:
          $ref: '#/components/schemas/Link'
        prev:
          $ref: '#/components/schemas/Link'
        next:
          $ref: '#/components/schemas/Link'
        last:
          $ref: '#/components/schemas/Link'
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Ressource non trouvée"
        code:
          type: string
          example: "NOT_FOUND"
        details:
          type: object